Refactored DestinationPage, added typing to DestinationTab
