+++ src/types/interfaces.ts
@@ -1,0 +1,8 @@
+export interface DestinationTabInterface {
+  src: string;
+  alt: string;
+  name: string;
+  description: string;
+  distance: string;
+  travel: string;
+}
